Fix numerical inaccuracies in set method
In the set method, adding a delta to update internal nodes according to
the new leaf value leads to numerical inaccuracies. Thus, the tree is
no longer a sum tree and therefore the sample method can return an
index greater than the current size of the tree. In other words, the
sample method can return the index of a non-existing node.

Instead of propagating a delta, the parents of the updated nodes are
recomputed with the new value. The method is slightly more expensive
but ensures that the tree is always a sum tree.